refactor(storage): repoint existing table_existence importers to introspection
Ref polylogue-48h

Mechanical import-path update for the 25 modules that already imported
table_exists/table_exists_async from the (now renamed) storage.table_existence
module, following the rename in the previous commit. No behavior change.
