+++ polylogue/cli/read_views/streaming_markdown.py
@@ -12,7 +12,7 @@ from polylogue.rendering.block_models import RenderableBlock
 from polylogue.rendering.blocks import has_structured_blocks, render_blocks_markdown
 from polylogue.rendering.core_markdown import format_message_text
 from polylogue.rendering.core_messages import normalize_render_timestamp
-from polylogue.storage.table_existence import table_exists as _table_exists
+from polylogue.storage.introspection import table_exists as _table_exists
 
 
 def stream_exact_session_markdown(
