+++ polylogue/storage/blob_publication.py
@@ -13,7 +13,7 @@ from typing import IO, BinaryIO
 from uuid import uuid4
 
 from polylogue.storage.blob_store import BlobStore, Heartbeat, PreparedBlob
-from polylogue.storage.table_existence import table_exists as _table_exists
+from polylogue.storage.introspection import table_exists as _table_exists
 
 
 @dataclass(frozen=True, slots=True)
