+++ polylogue/storage/blob_repair.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from polylogue.config import Config
 from polylogue.logging import get_logger
-from polylogue.storage.table_existence import table_exists as _table_exists
+from polylogue.storage.introspection import table_exists as _table_exists
 
 logger = get_logger(__name__)
