+++ polylogue/storage/search/runtime.py
@@ -9,12 +9,12 @@ from pathlib import Path
 
 from polylogue.core.errors import DatabaseError
 from polylogue.storage.fts.fts_lifecycle import check_fts_readiness, message_fts_search_readiness_sync
+from polylogue.storage.introspection import table_exists as _table_exists
 from polylogue.storage.search.cache import SearchCacheKey
 from polylogue.storage.search.models import SearchHit, SearchResult
 from polylogue.storage.search.query_builders import build_ranked_session_search_query, session_web_url
 from polylogue.storage.search.query_support import normalize_fts5_query, sort_key_to_iso
 from polylogue.storage.sqlite.connection import open_read_connection
-from polylogue.storage.table_existence import table_exists as _table_exists
 
 
 @lru_cache(maxsize=128)
