+++ polylogue/storage/sqlite/archive_tiers/session_annotations_write.py
@@ -24,7 +24,7 @@ import json
 import sqlite3
 from dataclasses import dataclass
 
-from polylogue.storage.table_existence import table_exists as _table_exists
+from polylogue.storage.introspection import table_exists as _table_exists
 
 
 @dataclass(frozen=True, slots=True)
