+++ polylogue/storage/sqlite/archive_tiers/user_audit.py
@@ -6,8 +6,8 @@ import sqlite3
 from dataclasses import dataclass
 
 from polylogue.core.json import JSONDocument, json_document
+from polylogue.storage.introspection import table_exists as _table_exists
 from polylogue.storage.sqlite.archive_tiers.user_write import AssertionKind
-from polylogue.storage.table_existence import table_exists as _table_exists
 
 _ASSERTION_BACKED_SURFACES: dict[str, AssertionKind] = {
     "marks": AssertionKind.MARK,
